MINIFICPP-1323 Move encryption functionality outside ConfigFile
